requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aioschedule==0.5.2
aiosqlite==0.17.0
async-timeout==3.0.1
asyncpg==0.23.0
asyncudp==0.4.0
attrs==21.2.0
Automat==20.2.0
bencode.py==4.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
greenlet==1.1.0
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
httptools==0.2.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
Jinja2==2.10.1
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pysqlite3==0.4.6
queuelib==1.6.1
requests==2.25.1
sanic==21.3.4
sanic-routing==0.6.2
schedule==1.1.0
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.18
tmdb3==0.7.2
tmdbsimple==2.8.0
torrent-tracker-scraper==4.1.0
tracker-scraper==1.1.0
Twisted==21.2.0
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvloop==0.15.2
w3lib==1.22.0
websockets==8.1
yarl==1.6.3
zope.interface==5.4.0
